app/requirements.txt
scikit-learn==1.5.2
pandas==2.2.3
category-encoders==2.6.4
pydantic==2.6.1
pydantic_core==2.16.2
fastapi==0.109.2
typing_extensions==4.9.0
uvicorn==0.27.1
requests==2.31.0
PyYAML==6.0.1
ipykernel==6.29.4
boto3==1.35.86
mlflow==2.19.0